+++ Changelog.md
@@ -379,3 +379,5 @@
 *	v1.27.67
 	*	Forced contiguous byte arrays in c3DSkit LZ11 compression to fix segfaults
 	*	Fixed a stupid regression in unpack.BCSTM
+*	v1.27.68
+	*	Fixed c3DSkit/graphics.c for little endian textures